requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-Cors==5.0.0
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.1
obs-websocket-py==1.0
obsws-python==1.7.0
tomli==2.0.2
websocket-client==1.8.0
Werkzeug==3.0.4
zipp==3.20.2
